Scripts update

+++ scripts/scripts/mesh-config.sh
@@ -1,18 +1,31 @@
 #!/bin/bash
-DEV=$1
+ENV=$1
 
 greymatter version
 
-source ./scripts/scripts/mesh-env.sh $DEV
-kubectl cluster-info
-listener=$(lsof -t -i:10080)
+source ./scripts/scripts/mesh-env.sh
 
+if [[ $ENV == "k3d" && "$(kubectl config current-context)" != "greymatter" ]]; then
+    echo "⛔️⛔️⛔️⛔️⛔ You are about to apply mesh configs to non-k3d environment ⛔️⛔️⛔️⛔️⛔️"
+    exit 1
+fi
+
+if [[ $ENV == "prod" && "$(kubectl config current-context)" = arn:aws:eks:*  ]]; then
+    echo "🍀🍀🍀🍀🍀 Just so you know, you are changing production 🍀🍀🍀🍀🍀"
+    source ./scripts/scripts/credentials.sh
+    if [ -z "$ClientId" ] || [ -z "$ClientSecret" ]; then
+        echo "⛔️ You do not have OIDC client ID and secret set in your credentials.sh"
+        exit 1
+    fi
+fi
+
+listener=$(lsof -t -i:10080)
 if [ ! -z "$listener" ]; then
     echo "Killing a process (pid $listener) using port 10080"
     kill $listener
 fi
 
-kubectl -n fabric wait --for=condition=Ready pod/api-0  --timeout=300s
+kubectl -n fabric wait --for=condition=Ready pod/api-0 --timeout=300s
 
 kubectl port-forward api-0 -n fabric 10080:10080 &
 sleep 10
@@ -44,21 +57,18 @@ create_or_update() {
 delay=0.01
 objects="domains clusters listeners proxies rules routes"
 meshfolders=(mesh/edge mesh/data/data mesh/data/jwt mesh/sense/catalog mesh/sense/dashboard mesh/sense/objectives mesh/sense/prometheus mesh/kibana mesh/website)
-for meshfolder in "${meshfolders[@]}"
-do
+for meshfolder in "${meshfolders[@]}"; do
     cd $meshfolder
-    for folder in $objects
-    do
+    for folder in $objects; do
         echo "Found folder: $meshfolder/$folder"
-        for file in $folder/*
-        do
+        for file in $folder/*; do
             object="${folder%?}"
             if [[ $object == "proxie" ]]; then object="proxy"; fi
             if [[ $object == "rule" ]]; then object="shared_rules"; fi
-            if [[ $object == "listener" && "$DEV" =~ ^([yY])$ ]]; then
+            if [[ $object == "listener" && $ENV == "k3d" ]]; then
                 value=$(<$file)
                 value=$(jq '.http_filters.gm_observables.useKafka = false' <<<"$value")
-                echo "$value" > /tmp/listener.json
+                echo "$value" >/tmp/listener.json
                 create_or_update listener /tmp/listener.json
                 sleep $delay
             else
@@ -72,44 +82,15 @@ do
 done
 
 # Overwrite
-if [[ "$DEV" =~ ^([yY])$ ]]; then
+if [[ $ENV == "k3d" ]]; then
     create_or_update domain scripts/resources/mesh.edge.domains.login.json
     create_or_update listener scripts/resources/mesh.edge.listener.login.json
     create_or_update listener scripts/resources/mesh.edge.listener.ingress.json
 fi
 
-if [[ ! "$(kubectl get pods -n apis)" ]]
-then
-    exit 0
+if [[ $ENV == "prod" ]]; then
+    cp mesh/edge/listeners/login.json /tmp/login.json
+    sed -i '' "s/CLIENT_ID_REDACTED/${ClientId}/g" /tmp/login.json
+    sed -i '' "s/CLIENT_SECRET_REDACTED/${ClientSecret}/g" /tmp/login.json
+    create_or_update listener /tmp/login.json
 fi
-
-kubectl port-forward deployment/catalog -n sense 10081:10080 &
-
-for meshfolder in apis/*
-do
-    if [ -d "$meshfolder" ]
-    then
-	for folder in $objects
-	do
-	    echo "Found folder: $meshfolder/mesh/$folder"
-	    for file in $meshfolder/mesh/$folder/*
-	    do
-		object="${folder%?}"
-		if [[ $object == "proxie" ]]; then object="proxy"; fi
-		if [[ $object == "rule" ]]; then object="shared_rules"; fi
-		if [[ $object == "listener" && "$DEV" =~ ^([yY])$ ]]; then
-                    value=$(<$file)
-                    value=$(jq '.http_filters.gm_observables.useKafka = false' <<<"$value")
-                    echo "$value" > /tmp/listener.json
-                    create_or_update listener /tmp/listener.json
-                    sleep $delay
-		else
-                    echo "applying $file"
-                    create_or_update $object $file
-                    sleep $delay
-		fi
-	    done
-	done
-	curl -XPOST http://localhost:10081/clusters -d "@$meshfolder/mesh/catalog.${meshfolder##*/}.json"
-    fi
-done
